feat(LottoGame): 수익금 형 int -> long 수정

+++ src/main/java/lotto/service/LottoGame.java
@@ -42,7 +42,7 @@ public class LottoGame {
 
     public double calculateRateOfReturn(PurchaseAmount purchaseAmount, ResultSheet resultSheet) {
         int amount = purchaseAmount.getAmount();
-        int proceeds = resultSheet.calculateProceeds();
+        long proceeds = resultSheet.calculateProceeds();
         return ((double) proceeds / amount) * 100;
     }
 }
